Implement spin animation for wheel
Create a rotation animation for the wheel view, so when user clicks on
the Spin button, the wheel is spun a random amount of degrees along its
axis and stays in the position it stops in.
Spin button is disabled while the wheel is still spinning, in order
to prevent the animation from jumping around.
The amount of degrees the wheel is randomly spun and time it is spun for
will likely be adjusted later when more is implemented.