Create WheelChoice object and database
Create object for wheel choices, which currently contains data for
choice's text, color, and times landed on.
Eventually user will be able to enter their own wheel choices, but for
now it is pre-populated with data when the db is created.
Use Room for database to keep code simple. Current database retrieval in
activity, but will move it out of there soon.
Alternative option was to store wheel choices in SharedPreferences, but
Room/SQLite database gives more flexibility in future if want to add
further data for a wheel choice like times landed on or lists of wheel
choices.