Refactor MainActivity into MVP pattern
Rename MainActivity to WheelActivity, as InputActivity will be built in
near future.
Refactor code in WheelActivity, putting business logic in
WheelPresenter, creating new custom WheelView class for code involving
drawing the wheel, and keeping ui logic in WheelActivity.
Change coroutines to more suitable ones rather than GlobalScope.
In WheelPresenter, create logic to determine which choice the wheel
lands on and return that info to the WheelActivity to display. This
logic may be moved out of this class in near future.