Create InputActivity
Remove prepopulate data in WheelChoiceDB and create InputActivity (along
with InputPresenter) for user to enter in up to 5 WheelChoices to later
display in the WheelActivity.
If enough choices are entered (at least 2), user is allowed to go to
WheelActivity. Choices are editable at any time from the InputActivity.
Still need to add option for user to select colors.